Fixing displayed text in labels, dialogs. Moving enum converters in individual files.
