+++ UserDefinedArgumentTypeConverter.cs
@@ -1,0 +1,8 @@
+﻿using SSMSObjectExplorerMenu.enums;
+
+namespace SSMSObjectExplorerMenu
+{
+    public class UserDefinedArgumentTypeConverter : EnumConverter<UserDefinedParameterType>
+    {
+    }
+}
